Refactor TabmanViewController extensions
