+++ Sources/Tabman/TabmanViewController.swift
@@ -19,9 +19,9 @@ open class TabmanViewController: PageboyViewController, PageboyViewControllerDel
     /// The instance managed Tabman bar.
     internal fileprivate(set) var tabmanBar: TabmanBar?
     /// Currently attached TabmanBar if it exists.
-    internal fileprivate(set) var attachedTabmanBar: TabmanBar?
+    internal var attachedTabmanBar: TabmanBar?
     /// The view that is currently being used to embed the instance managed TabmanBar.
-    internal fileprivate(set) var embeddingView: UIView?
+    internal var embeddingView: UIView?
     
     /// Returns the active bar, prefers attachedTabmanBar if available.
     fileprivate var activeTabmanBar: TabmanBar? {
@@ -53,6 +53,12 @@ open class TabmanViewController: PageboyViewController, PageboyViewControllerDel
         self.updateBar(withLocation: self.bar.location)
     }
     
+    open override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        
+        self.insetChildViewControllerIfNeeded(self.currentViewController)
+    }
+    
     open override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
         super.viewWillTransition(to: size, with: coordinator)
         let bounds = CGRect(x: 0.0, y: 0.0, width: size.width, height: size.height)
@@ -70,6 +76,8 @@ open class TabmanViewController: PageboyViewController, PageboyViewControllerDel
                                       willScrollToPageAtIndex index: Int,
                                       direction: PageboyViewController.NavigationDirection,
                                       animated: Bool) {
+        self.insetChildViewControllerIfNeeded(self.viewControllers?[index])
+        
         if animated {
             UIView.animate(withDuration: 0.3, animations: {
                 self.activeTabmanBar?.updatePosition(CGFloat(index), direction: direction)
@@ -214,83 +222,13 @@ internal extension TabmanViewController {
     }
 }
 
-// MARK: - TabmanBar attach/detachment
-public extension TabmanViewController {
+// MARK: - Child view controller layout
+internal extension TabmanViewController {
     
-    /// Attach a TabmanBar that is somewhere in the view hierarchy.
-    /// This will replace the TabmanViewController managed instance.
-    ///
-    /// - Parameter bar: The bar to attach.
-    public func attach(bar: TabmanBar) {
-        guard self.attachedTabmanBar == nil else { return }
+    func insetChildViewControllerIfNeeded(_ childViewController: UIViewController?) {
+        guard let childViewController = childViewController else { return }
         
-        self.tabmanBar?.isHidden = true
-        self.reloadRequiredBarInsets()
-        
-        // hook up new bar
-        bar.dataSource = self
-        bar.delegate = self
-        bar.transitionStore = self.barTransitionStore
-        if let appearance = self.bar.appearance {
-            bar.appearance = appearance
-        }
-        bar.isHidden = true
-        self.attachedTabmanBar = bar
-        
-        bar.reloadData()
-    }
-    
-    
-    /// Detach a currently attached external TabmanBar.
-    /// This will reinstate the TabmanViewController managed instance.
-    ///
-    /// - Returns: The detached bar.
-    @discardableResult public func detachAttachedBar() -> TabmanBar? {
-        guard let bar = self.attachedTabmanBar else { return nil }
-        guard self.attachedTabmanBar === bar else { return nil }
-        
-        bar.dataSource = nil
-        bar.delegate = nil
-        bar.transitionStore = nil
-        bar.isHidden = false
-        self.attachedTabmanBar = nil
-        
-        self.tabmanBar?.reloadData()
-        self.view.layoutIfNeeded()
-
-        self.reloadRequiredBarInsets()
-        
-        return bar
-    }
-    
-    /// Embed the TabmanBar in an external view.
-    /// This will add the bar to the specified view, and pin the bar edges to the view edges.
-    ///
-    /// - Parameter view: The view to embed the bar in.
-    public func embedBar(inView view: UIView) {
-        guard let bar = self.tabmanBar else { return }
-        guard self.embeddingView == nil || view === self.embeddingView else { return }
-
-        self.embeddingView = view
-        
-        bar.removeFromSuperview()
-        view.addSubview(bar)
-        bar.autoPinEdgesToSuperviewEdges()
-        self.reloadRequiredBarInsets()
-
-        view.layoutIfNeeded()
-    }
-    
-    /// Disembed the TabmanBar from an external view if it is currently embedded.
-    public func disembedBar() {
-        guard let bar = self.tabmanBar else { return }
-        guard self.embeddingView != nil else { return }
-        
-        bar.removeFromSuperview()
-        self.embeddingView = nil
-        
-        self.updateBar(withLocation: self.bar.location)
-        self.reloadRequiredBarInsets()
+        print(childViewController)
     }
 }
 
@@ -337,7 +275,7 @@ extension TabmanViewController: TabmanBarConfigDelegate {
     
     func config(_ config: TabmanBarConfig, didUpdateItems items: [TabmanBarItem]?) {
         self.activeTabmanBar?.reloadData()
-        
+
         self.view.layoutIfNeeded()
         self.reloadRequiredBarInsets()
     }
